Ensure redis configuration is actually changed for each instance.
The redis connection is memoized as a class-instance variable on `Sidekiq`.
This means that setting the redis configuration has no effect after the first
time within the same process.

It looks as if the NewRelic gem runs all plugins in the same process, so we
need to reset the redis configuration somehow. I can't see a clean way of
doing this with the existing `Sidekiq` API, so this hack will have to do
for now.